requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
mysqlclient==2.2.7
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
PyYAML==6.0.2
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
uritemplate==4.1.1
